re_exec_indentf,re_indentf: silence warnings
Pass the right types to printf.

For re_exec_indentf(), really the type of the depth arg should be changed
so that it and the depth var are consistent throughout regexec.c, but
that's probably something for post-5.24.

+++ regcomp.c
@@ -933,7 +933,7 @@ Perl_re_indentf(pTHX_ const char *fmt, U32 depth, ...)
     PerlIO *f= Perl_debug_log;
     PERL_ARGS_ASSERT_RE_INDENTF;
     va_start(ap, depth);
-    PerlIO_printf(f, "%*s", ( depth % 20 ) * 2, "");
+    PerlIO_printf(f, "%*s", ( (int)depth % 20 ) * 2, "");
     result = PerlIO_vprintf(f, fmt, ap);
     va_end(ap);
     return result;
